--- a/package.json
+++ b/package.json
@@ -20,7 +20,8 @@
 		"node": ">=18"
 	},
 	"scripts": {
-		"test": "xo && ava && tsd"
+		"test": "xo && ava && tsd",
+		"benchmark": "node benchmark/index.js"
 	},
 	"files": [
 		"index.js",