Fixes #10

Apparently, creating a `vm` context is very expensive. It now reuses the context. This should in theory be safe as JavaScript is single-threaded, so two functions cannot execute, for example, `firstMatch` at the same time. But it does make me think whether there's anything I'm missing.

// @fityannugroho


---

```
❯ nr benchmark

> super-regex@0.2.0 benchmark
> node benchmark/index.js

str.matchAll: 0.236ms
matches: 2.355ms
matches-t500: 16.928ms
str.match: 0.149ms
firstMatch: 2.334ms
firstMatch-t500: 22.152ms

	Count chars: 23749
	Count rows: 405
	Count matches: 363 363 363 363 363 363
```